Added the projects stable files.
This component will be improved on later. But for now it is working
minimally in terms of what it can do. There will be some more attributes
to be added and this will help expand the functionality of this button.

It will also have RTL support in the future iterations.